Created README.md
Hi Ron! 

I was just browsing GitHub and found this apk. I tried it on my phone but I could not understand it completely. 
What I could see was I was sitting inside a car and that was amazing! 

Could you please add some description to this README.md file. It will be of great help to understand the working of the application. 

Keep up the good work, cheers! :D